+++ hata/discord/user/guild_profile/guild_profile.py
@@ -8,11 +8,13 @@ from ...core import ROLES
 from ...utils import DISCORD_EPOCH_START
 
 from .fields import (
-    parse_boosts_since, parse_joined_at, parse_nick, parse_pending, parse_role_ids, parse_timed_out_until,
-    put_boosts_since_into, put_joined_at_into, put_nick_into, put_pending_into, put_role_ids_into,
-    put_timed_out_until_into, validate_boosts_since, validate_joined_at, validate_nick, validate_pending,
-    validate_role_ids, validate_timed_out_until
+    parse_boosts_since, parse_flags, parse_joined_at, parse_nick, parse_pending, parse_role_ids, parse_timed_out_until,
+    put_boosts_since_into, put_flags_into, put_joined_at_into, put_nick_into, put_pending_into, put_role_ids_into,
+    put_timed_out_until_into, validate_boosts_since, validate_flags, validate_joined_at, validate_nick,
+    validate_pending, validate_role_ids, validate_timed_out_until
 )
+from .flags import GuildProfileFlag
+
 
 create_partial_role_from_id = include('create_partial_role_from_id')
 
@@ -33,6 +35,9 @@ class GuildProfile(metaclass = Slotted):
         Since when the user uses it's Nitro to boost the respective guild.
         If the user does not boost the guild then is set as `None`.
     
+    flags : ``GuildProfileFlag``
+        The guild profile's flags.
+    
     joined_at : `None`, `datetime`
         The date, since the user is the member of the guild.
         If this field was not included with the initial data, then it is set to `None`.
@@ -50,21 +55,27 @@ class GuildProfile(metaclass = Slotted):
     timed_out_until : `None`, `datetime`
         Till when the user is timed out, and cannot interact with the guild.
     """
-    __slots__ = ('boosts_since', 'joined_at', 'nick', 'pending', 'role_ids', 'timed_out_until')
+    __slots__ = ('boosts_since', 'flags', 'joined_at', 'nick', 'pending', 'role_ids', 'timed_out_until')
     
     avatar = IconSlot('avatar', 'avatar', None, None)
     
-    def __new__(cls, **keyword_parameters):
+    def __new__(
+        cls,
+        *,
+        avatar = ...,
+        boosts_since = ...,
+        flags = ...,
+        joined_at = ...,
+        nick = ...,
+        pending = ...,
+        role_ids = ...,
+        timed_out_until = ...,
+    ):
         """
         Creates a new guild profile instance from the given parameters.
         
         Parameters
         ----------
-        **keyword_parameters : Keyword parameters
-            Keyword parameters defining how should the guild profile's parameters be set.
-        
-        Other Parameters
-        ----------------
         avatar : `None`, ``Icon``, `str`, Optional (Keyword only)
             The channel's avatar.
             
@@ -79,10 +90,13 @@ class GuildProfile(metaclass = Slotted):
             The channel's avatar's hash.
             
             > Mutually exclusive with the `avatar` parameter.
-            
+        
         boosts_since : `None`, `datetime`, Optional (Keyword only)
             Since when the user uses it's Nitro to boost the respective guild.
         
+        flags : `int`, ``GuildProfileFlag``, Optional (Keyword only)
+            The guild profile's flags.
+            
         joined_at : `None`, `datetime`, Optional (Keyword only)
             The date, since the user is the member of the guild.
         
@@ -106,8 +120,64 @@ class GuildProfile(metaclass = Slotted):
         ValueError
             - If a parameter's value is incorrect.
         """
-        self = cls._create_empty()
-        self._populate_from_keyword_parameters(keyword_parameters)
+        # avatar
+        if avatar is ...:
+            avatar = None
+        else:
+            avatar = cls.avatar.validate_icon(avatar, allow_data = True)
+        
+        # boosts_since
+        if boosts_since is ...:
+            boosts_since = None
+        else:
+            boosts_since = validate_boosts_since(boosts_since)
+        
+        # flags
+        if flags is ...:
+            flags = GuildProfileFlag()
+        else:
+            flags = validate_flags(flags)
+        
+        # joined_at
+        if joined_at is ...:
+            joined_at = None
+        else:
+            joined_at = validate_joined_at(joined_at)
+        
+        # nick
+        if nick is ...:
+            nick = None
+        else:
+            nick = validate_nick(nick)
+        
+        # pending
+        if pending is ...:
+            pending = False
+        else:
+            pending = validate_pending(pending)
+        
+        # role_ids
+        if role_ids is ...:
+            role_ids = None
+        else:
+            role_ids = validate_role_ids(role_ids)
+        
+        # timed_out_until
+        if timed_out_until is ...:
+            timed_out_until = None
+        else:
+            timed_out_until = validate_timed_out_until(timed_out_until)
+        
+        # Construct
+        self = object.__new__(cls)
+        self.avatar = avatar
+        self.boosts_since = boosts_since
+        self.flags = flags
+        self.joined_at = joined_at
+        self.nick = nick
+        self.pending = pending
+        self.role_ids = role_ids
+        self.timed_out_until = timed_out_until
         return self
     
     
@@ -127,6 +197,9 @@ class GuildProfile(metaclass = Slotted):
         boosts_since = self.boosts_since
         if (boosts_since is not None):
             hash_value ^= hash(boosts_since)
+        
+        # flags
+        hash_value ^= self.flags
         
         # joined_at
         joined_at = self.joined_at
@@ -167,6 +240,10 @@ class GuildProfile(metaclass = Slotted):
         
         # boosts_since
         if self.boosts_since != other.boosts_since:
+            return False
+        
+        # flags
+        if self.flags != other.flags:
             return False
         
         # joined_at
@@ -230,29 +307,16 @@ class GuildProfile(metaclass = Slotted):
         """
         data = {}
         
-        # avatar
         type(self).avatar.put_into(self.avatar, data, defaults, as_data = not include_internals)
+        put_nick_into(self.nick, data, defaults)
+        put_role_ids_into(self.role_ids, data, defaults)
+        put_timed_out_until_into(self.timed_out_until, data, defaults)
         
-        # boosts_since
         if include_internals:
             put_boosts_since_into(self.boosts_since, data, defaults)
-        
-        # joined_at
-        if include_internals:
+            put_flags_into(self.flags, data, defaults)
             put_joined_at_into(self.joined_at, data, defaults)
-        
-        # nick
-        put_nick_into(self.nick, data, defaults)
-        
-        # pending
-        if include_internals:
             put_pending_into(self.pending, data, defaults)
-        
-        # role_ids
-        put_role_ids_into(self.role_ids, data, defaults)
-        
-        # timed_out_until
-        put_timed_out_until_into(self.timed_out_until, data, defaults)
         
         return data
     
@@ -281,6 +345,7 @@ class GuildProfile(metaclass = Slotted):
         """
         self._set_avatar(data)
         self.boosts_since = parse_boosts_since(data)
+        self.flags = parse_flags(data)
         self.nick = parse_nick(data)
         self.pending = parse_pending(data)
         self.role_ids = parse_role_ids(data)
@@ -312,6 +377,8 @@ class GuildProfile(metaclass = Slotted):
         +-------------------+-------------------------------+
         | boosts_since      | `None`, `datetime`            |
         +-------------------+-------------------------------+
+        | flags             | `None`, ``GuildProfileFlags`` |
+        +-------------------+-------------------------------+
         | nick              | `None`, `str`                 |
         +-------------------+-------------------------------+
         | pending           | `bool`                        |
@@ -332,6 +399,11 @@ class GuildProfile(metaclass = Slotted):
             old_attributes['boosts_since'] = self.boosts_since
             self.boosts_since = boosts_since
         
+        flags = parse_flags(data)
+        if self.flags != flags:
+            old_attributes['flags'] = self.flags
+            self.flags = flags
+            
         # nick
         nick = parse_nick(data)
         if self.nick != nick:
@@ -371,6 +443,7 @@ class GuildProfile(metaclass = Slotted):
         new.avatar_hash = self.avatar_hash
         new.avatar_type = self.avatar_type
         new.boosts_since = self.boosts_since
+        new.flags = self.flags
         new.joined_at = self.joined_at
         new.nick = self.nick
         new.pending = self.pending
@@ -382,17 +455,23 @@ class GuildProfile(metaclass = Slotted):
         return new
     
     
-    def copy_with(self, **keyword_parameters):
+    def copy_with(
+        self, 
+        *,
+        avatar = ...,
+        boosts_since = ...,
+        flags = ...,
+        joined_at = ...,
+        nick = ...,
+        pending = ...,
+        role_ids = ...,
+        timed_out_until = ...,
+    ):
         """
         Copies the guild profile and modifies the defined the defined fields of it.
         
         Parameters
         ----------
-        **keyword_parameters : Keyword parameters
-            Keyword parameters defining how should the guild profile's parameters be set.
-        
-        Other Parameters
-        ----------------
         avatar : `None`, ``Icon``, `str`, Optional (Keyword only)
             The channel's avatar.
             
@@ -410,6 +489,9 @@ class GuildProfile(metaclass = Slotted):
             
         boosts_since : `None`, `datetime`, Optional (Keyword only)
             Since when the user uses it's Nitro to boost the respective guild.
+        
+        flags : `int`, ``GuildProfileFlag``, Optional (Keyword only)
+            The guild profile's flags.
         
         joined_at : `None`, `datetime`, Optional (Keyword only)
             The date, since the user is the member of the guild.
@@ -438,8 +520,66 @@ class GuildProfile(metaclass = Slotted):
         ValueError
             - If a parameter's value is incorrect.
         """
-        new = self.copy()
-        new._populate_from_keyword_parameters(keyword_parameters)
+        # avatar
+        if avatar is ...:
+            avatar = self.avatar
+        else:
+            avatar = type(self).avatar.validate_icon(avatar, allow_data = True)
+        
+        # boosts_since
+        if boosts_since is ...:
+            boosts_since = self.boosts_since
+        else:
+            boosts_since = validate_boosts_since(boosts_since)
+        
+        # flags
+        if flags is ...:
+            flags = self.flags
+        else:
+            flags = validate_flags(flags)
+        
+        # joined_at
+        if joined_at is ...:
+            joined_at = self.joined_at
+        else:
+            joined_at = validate_joined_at(joined_at)
+        
+        # nick
+        if nick is ...:
+            nick = self.nick
+        else:
+            nick = validate_nick(nick)
+        
+        # pending
+        if pending is ...:
+            pending = False
+        else:
+            pending = validate_pending(pending)
+        
+        # role_ids
+        if role_ids is ...:
+            role_ids = self.role_ids
+            if (role_ids is not None):
+                role_ids = (*role_ids,)
+        else:
+            role_ids = validate_role_ids(role_ids)
+        
+        # timed_out_until
+        if timed_out_until is ...:
+            timed_out_until = self.timed_out_until
+        else:
+            timed_out_until = validate_timed_out_until(timed_out_until)
+        
+        # Construct
+        new = object.__new__(type(self))
+        new.avatar = avatar
+        new.boosts_since = boosts_since
+        new.flags = flags
+        new.joined_at = joined_at
+        new.nick = nick
+        new.pending = pending
+        new.role_ids = role_ids
+        new.timed_out_until = timed_out_until
         return new
     
     
@@ -456,88 +596,13 @@ class GuildProfile(metaclass = Slotted):
         self.avatar_hash = 0
         self.avatar_type = IconType.none
         self.boosts_since = None
+        self.flags = GuildProfileFlag()
         self.joined_at = None
         self.nick = None
         self.pending = False
         self.role_ids = None
         self.timed_out_until = None
         return self
-    
-
-    def _populate_from_keyword_parameters(self, keyword_parameters):
-        """
-        Sets the guild profile's attributes based on the given values.
-        
-        Parameters
-        ----------
-        keyword_parameters : `dict` of (`str`, `Any`) items
-            Keyword parameters defining how should the guild profile's parameters be set.
-        
-        Raises
-        ------
-        TypeError
-            - Extra or unused parameters.
-        """
-        if not keyword_parameters:
-            return
-        
-        # avatar
-        avatar = type(self).avatar.parse_from_keyword_parameters(keyword_parameters, allow_data = False)
-        if (avatar is not None):
-            self.avatar = avatar
-        
-        # boosts_since
-        try:
-            boosts_since = keyword_parameters.pop('boosts_since')
-        except KeyError:
-            pass
-        else:
-            self.boosts_since = validate_boosts_since(boosts_since)
-        
-        # joined_at
-        try:
-            joined_at = keyword_parameters.pop('joined_at')
-        except KeyError:
-            pass
-        else:
-            self.joined_at = validate_joined_at(joined_at)
-        
-        # nick
-        try:
-            nick = keyword_parameters.pop('nick')
-        except KeyError:
-            pass
-        else:
-            self.nick = validate_nick(nick)
-        
-        # pending
-        try:
-            pending = keyword_parameters.pop('pending')
-        except KeyError:
-            pass
-        else:
-            self.pending = validate_pending(pending)
-        
-        # role_ids
-        try:
-            role_ids = keyword_parameters.pop('role_ids')
-        except KeyError:
-            pass
-        else:
-            self.role_ids = validate_role_ids(role_ids)
-        
-        # timed_out_until
-        try:
-            timed_out_until = keyword_parameters.pop('timed_out_until')
-        except KeyError:
-            pass
-        else:
-            self.timed_out_until = validate_timed_out_until(timed_out_until)
-        
-        if keyword_parameters:
-            raise TypeError(
-                f'Extra or unused keyword parameters: {keyword_parameters!r}.'
-            )
     
     
     def get_top_role(self, default = None):