+++ hata/discord/user/guild_profile/tests/test__put_flags_into.py
@@ -1,0 +1,17 @@
+import vampytest
+
+from ..fields import put_flags_into
+from ..flags import GuildProfileFlag
+
+
+def test__put_flags_into():
+    """
+    Tests whether ``put_flags_into`` is working as intended.
+    """
+    for input_, defaults, expected_output in (
+        (GuildProfileFlag(0), False, {}),
+        (GuildProfileFlag(0), True, {'flags': 0}),
+        (GuildProfileFlag(1), False, {'flags': 1}),
+    ):
+        data = put_flags_into(input_, {}, defaults)
+        vampytest.assert_eq(data, expected_output)